backend/requirements.txt
# Web/API
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
joblib==1.4.2
scikit-learn==1.5.2
supabase==2.6.0
psycopg2==2.9.10


# NLP / parsing
scikit-learn==1.5.2
nltk==3.9.1
pdfminer.six==20231228
beautifulsoup4==4.12.3
lxml==5.3.0

# E-mail (IMAP) e parsing RFC822
imap-tools==1.5.0

# Utilitários
python-dotenv==1.0.1

# --- Opcionais (IA externa/zero-shot/LLM) ---
# transformers==4.44.2
# torch==2.4.0
# openai==1.43.0
